+++ src/ui/RealmBrowser/AddClassModal/index.tsx
@@ -40,7 +40,7 @@ const initialState: IAddClassModalState = {
   nameIsValid: true,
   primaryKey: false,
   primaryKeyName: '',
-  primaryKeyType: 'string',
+  primaryKeyType: 'object id',
 };
 
 class AddClassModalContainer extends React.Component<
@@ -87,7 +87,7 @@ class AddClassModalContainer extends React.Component<
   };
 
   private preparePrimaryKeyName = (primaryKeyName: string) =>
-    primaryKeyName === '' ? 'uuid' : primaryKeyName;
+    primaryKeyName === '' ? '_id' : primaryKeyName;
 
   private getSchema = (): Realm.ObjectSchema => {
     const { name, primaryKey, primaryKeyType } = this.state;