+++ RELEASENOTES.md
@@ -7,7 +7,9 @@
 
 ### Enhancements
 
+- Added support for creating & generating `ObjectId` when creating objects. ([#1291](https://github.com/realm/realm-studio/pull/1291))
 - Added support for editing existing `ObjectId` values. ([#1290](https://github.com/realm/realm-studio/pull/1290))
+- Primary keys in Class/Schema creation now defaults to an `ObjectId` property named `_id`.
 
 ### Fixed